astro_reddit/requirements.txt
praw
pandas
nltk
google-cloud-storage
google-cloud-secret-manager
google-auth
requests